Add Chart for backup-restore-operator
This commit contains the serviceaccount for the operator and clusterrolebinding binding it to cluster-admin role.
The serviceaccount gets cluster-admin role because this chart's controllers need permissions to CRUD any resource.
Also the controller can get, upload and delete backups to s3, so only the cluster-admin role should be able to do this