Fix remove_timestamp_from_diff in generate-patch
This commit resolves two issues that are found when working with charts:
- On a failed sed command for removing the timestamp, the patch also
fails to regenerate the CRD package. Since removing the timestamps on the patch
before or after generating the CRD package makes no difference, we can just
move the `remove_timestamp_from_diff` command to the end of the for loop.
- The sed command would fail depending on what type of sed you were using.
The easiest fix is to provide a backup file extension name like `.bak` and just
delete the file after generating it to have it work with both types of sed.

Related comment for second issue: https://github.com/rancher/dev-charts/pull/38#discussion_r458442691