Move revert_crd_changes into a new script
Before, the logic for reverting a CRD chart was located in `generate-patch` since it was the only script that required these changes. Now that the same logic is required on a `make clean` in order to support reverting just the generateCRDChart changes from local charts, this commit moves that logic out to its own script `clean-crds`.

Based on discussion in https://github.com/rancher/charts/pull/607#discussion_r483288658.