Update scripts to roll back optional flag changes
This commit makes a change to the `clean`, `generate-patch`, and `validate` due to a series of edge cases that were introduced from not properly rolling back changes from optional flags that were introduced in `prepare` during intermediary steps of each of these scripts.

In general, the prepare script performs the following changes:
1) Pull from upstream (or keep a local version)
2) Prepare subcharts
3) Apply patch
4) Optional: generateCRDChart changes
5) Optional: providesGVR changes

To avoid patching on the optional flags added to the chart (since those resources shouldn't get tracked in a patch), we normally reverted change 4 and 5 in the `generate-patch` script in order to bring the chart back to a state where we can compute a diff of the patch in step 3. This was done by calling the `revert_crd` function. However, this behavior is now also needed to ensure that Git is not dirty on a `validate` or to bring a chart back to its normal state on a `clean` so it has now been split into the `revert-crds` script.