Replace yq in place commands with awk replaces
Until yq has a resolution that allows us to cleanly deal with inplace insertion, we need to provide a solution in the scripts that makes it possible for developers to be able to run the scripts without encountering random changes associated with Chart.yaml changes. This commit replaces all our inplace modificaitons using yq with helper functions that effectively do the same YAML manipulation using awk.