Split up CRDs based on ConfigMap 1mb limit
This commit changes the chart template to create a new ConfigMap if a single ConfigMap has over 1mb in it.

The previous behavior was to use one ConfigMap for all CRDs, which does not work for Rancher Monitoring since it has exceeded that (see issue).

It only affects charts that use `assumeOwnershipOfCRDs`.

Related Issue: https://github.com/rancher/rancher/issues/29973