Add global seLinux toggle to rancher-logging
Add a new value to the rancher-logging chart to enable setting the
seLinuxOptions securityContext attribute for Logging resources and the
rke-aggregator DaemonSet. This is applied to any resource that needs to
read from /var/log/containers or /var/lib/docker/containers.

Without this patch, components in the rancher-logging chart mount and
attempt to read and sometimes write files on several hostPaths, and when
SELinux is enabled on the host, will fail because the containers are
unauthorized to do so. The added seLinuxOptions attribute in this patch
causes the fluentbit and rke-aggregator DaemonSets to run with the
SELinux context type `rke_logreader_t`, which is provided by the
rke2-linux package and enables the container to read files with the
label `container_log_t`, which includes files like
/var/log/containers/*.log, and files with type `rke_container_var_lib_t`
which the docker log files are relabeled to.

The logging aggregator will still fail to read and write log data from
/var/lib/rancher/logging, but this will be addressed in a different way.