Fix RKE2 aggregator DaemonSet for rancher-logging
Without this patch, the rancher-logging chart fails to install when RKE2
is included:

```
$ helm install rancher-logging rancher/rancher-logging --set additionalLoggingSources.rke2.enabled=true
Error: YAML parse error on rancher-logging/templates/loggings/rke2/daemonset.yaml: error converting YAML to JSON: yaml: line 25: did not find expected key
```

This is because the indentation being generated was invalid, which
itself was caused by the invalid use of `tolerations` and `nodeSelector`
under the `containers` object, when these are only Pod attributes. This
change fixes the indendation so that the end result is a valid pod spec
and the chart installs correctly.