Switch internalPort to 8443 instead of 10250
In upstream kube-prometheus-stack, the initial purpose of this change was to enable GKE private clusters to deploy Monitoring without adding any firewall rules, since GKE keeps the port open by default for kubelet.

However, as noted in https://github.com/rancher/rancher/issues/32201#issuecomment-822677439, this results in an error since Prometheus Operator is trying to listen / bind to a port that is already in use by kubelet on the host.

The fix for this issue is to revert to using port 8443 for Prometheus Operator, which was already the expectation in previous releases of Rancher Monitoring, so it is assumed that users who are upgrading within a GKE private cluster would already have opened this port within their firewall rules and new users must create the approporiate firewall rules for this use case.

Upstream commit that is being partially reverted: https://github.com/prometheus-community/helm-charts/commit/164e84e852d5916627c6f02375b0f1c3c4c92290.