Fix alerting-drivers Chart.yaml
This is required due to the change released in v0.1.0 of the charts-build-scripts to switch the package used for YAML marshalling / unmarshalling.

Note: This does not need a packageVersion bump since it has already been bumped once for this release.
(cherry picked from commit e5415e48debbd68542956fc9a9a7317000650292)