Add rancher-version to rancher-operator
rancher-operator was merged into Rancher[1] and so the image and chart
won't be used for Rancher 2.6 going forward. This change adds a patch to
pin the rancher version to <2.6 so it won't be used in new installs and
won't be picked up by image-scanning. The change is made directly to the
chart because the package was removed in 143c0fec.

[1] https://github.com/rancher/rancher/pull/32170