Add rancher-version to rancher-operator
rancher-operator was merged into Rancher[1] and so the image and chart
won't be used for Rancher 2.6 going forward. This change adds a patch to
pin the rancher version to <2.6 so it won't be used in new installs and
won't be picked up by image-scanning.

[1] https://github.com/rancher/rancher/pull/32170

(cherry picked from commit 939f18a329fe8b7d33123a8066c612cc695f98a7)