Fix references to "exporter.COMPONENT.enabled" templates
Throughout the code, in order to ensure that we enable certain k8s-component-specific things on k3s servers even when the specific components are not enabled, we use the helper templates defined in the _helpers.tpl in order to trigger enabling specific alerting / recording rules for those components.

This commit fixes the references to ensure any .Values.COMPONENT.enabled calls are replaced with template calls to "exporter.COMPONENT.enabled".

It also introduces some minor changes, like deprecating the helper template used to identify whether the kubeletService is enabled, which were erroneously coded originally based on the current review of the code.