Introduce the capability to override the service selector for pushprox
This commit introduces the ability to supply an arbitrary selector for the Service object created by the chart which may not point to the PushProx client.

This setting is only possible if `.Values.clients.enabled` is set to false.

The expected use case of this feature is if both `.Values.clients.enabled` and `.Values.proxy.enabled` are set to false; in that case, this chart will only deploy a Service and ServiceMonitor resource, in which case modifying the Service will allow a user to use pushprox as a generic chart to set up an exporter that targets pods at a given `.Values.metricsPort` based on labels.