Ensure that Grafana datasources are loaded properly despite anonymous login
Related Upstream Issue: https://github.com/grafana/helm-charts/issues/981

By default Rancher Monitoring allows anonymous users to be able to log in and view dashboards, since it is assumed that users who have access to the Grafana endpoint are accessing it through Rancher, which means that they are already authenticated via the Rancher API. We do not leverage Grafana's basic authentication.

However, as noted in the upstream issue, the admin HTTP API specifically requires basic authentication to be enabled to be able to allow datasources to be reloaded.

To workaround this issue, this commit enables the default initContainer that was the previous behavior of the Grafana Helm chart to load datasources on initialization and skipping the reload URL.

An issue to track the proper fix for this has been filed in https://github.com/rancher/rancher/issues/39385.