Remove container label from container_network queries in Rancher dashboards and sum over irate
This commit fixes the queries calculated on the Rancher dashboards for network usage to mimic the way they are calculated in upstream.

This fixes an issue in the dashboards where Network panels would not show up since the container label was missing from the series scraped by Prometheus in an RKE 1.24 cluster.