Use older apiserver_request_duration_seconds when slo version is not available in <1.23.0-0 clusters
In upstream Kubernetes (specifically in this commit https://github.com/kubernetes/apiserver/commit/9144ea1240a23e7321ac9c9ea76d893e686ad8e1), starting k8s 1.23 a new version of the apiserver_request_duration_seconds metric was released called apiserver_request_slo_duration_seconds.

As noted in the pull request https://github.com/kubernetes/kubernetes/pull/105890, the fundamental difference between the newer and older metric is that the newer metric excludes the processing time spent by webhooks that the API server calls out to when configured as part of a MutatingWebhookConfiguration or ValidatingWebhookConfiguration.

However, since this metric was only introduced in k8s 1.23, this breaks the API Server dashboards in clusters <1.23.0 since that metric does not exist.

Therefore, to fix this issue in older clusters using the latest Monitoring chart, specific logic has been added to the chart to utilize the older version of the metric when the Kubernetes version is detected to be <1.23.