[dev-v2.7] Fix patch failures for newer versions of charts-build-scripts (#4014)
I am using admin privileges to merge this to unblock the release process of release-v2.7.14. 
The fleet team is aware of the situation and they had previously approved it but I needed to make extra changes later. 
Regarding epinio, there was no release of a new chart, only the packages folder so I am bypassing it as well since I could not get a response in time.