Remove Fleet 107.1.0+up0.13.4-rc.2
we initially bumped to 107.1.0 because there was an inbetween Rancher
versions Fleet release.

But when we used chart autobump to release the follow development
versions the minor version was decreased again, which went unnoticed
initally.

To prevent issues in the release script, we are removing this release.
It is an old development version anyway which final release is long out.